requirements.txt
django==3.1.6
pyyaml
psycopg2
Pillow
social-auth-app-django  
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
djangorestframework
celery==4.2.1
redis==2.10.6